requirements.txt
PyYAML>=6.0
Pillow>=9.0.0
reportlab>=3.6.0
PyPDF2>=3.0.0
python-dateutil>=2.8.0
pyinstaller>=5.0.0


